requirements.txt
django
djangorestframework
markdown
django-filter
bpython
httpie
django-extensions